chore: Add support for LaTeX math equations in MessageItem.jsx
